+++ doc/CHANGES.md
@@ -31,8 +31,9 @@ Change log
 
 ## v0.6.3-dev (upcoming changes)
 
-- fix [#1142](https://github.com/gridstack/gridstack.js/issues/1142) RemoveWidget() doesn't also trigger change events when it should
-- del `bower` since dead for a while now (https://snyk.io/blog/bower-is-dead/)
+- fix [#1142](https://github.com/gridstack/gridstack.js/issues/1142) add/remove widget will also trigger change events when it should.
+- optimized `change` callback to save original x,y,w,h values and only call those that changed [1148](https://github.com/gridstack/gridstack.js/pull/1148)
+- delete `bower` since [dead](https://snyk.io/blog/bower-is-dead) for a while now
 
 ## v0.6.3 (2020-02-05)